Convert computeProductsWithPrices to TypeScript
